weather_gui/requirements.txt
customtkinter==5.2.2
pillow==11.0.0
requests==2.32.3
